refactor: move unit of work to application layer instead of domain layer This has the implication of requiring a rethinking of the repository implementation which was previously expecting a transaction object. This violates the domain layer because db transaction objects shouldn't be a thing in the business logic.
The tricky part was figuring out how to obtain the domain service object in the application layer. The unit of work provides a repository provider which instantiates a repository with tx object. At the same time, the domain service object expects a repository object to be injected on instantiation. But domain services were instantiated in composition root.

Perhaps in the future I should reconsider if domain services should be instantiated in composition root.